requirements.txt
astroid==1.4.4
boto==2.38.0
colorama==0.3.6
Django==1.8
django-extensions==1.6.1
django-secure==1.0.1
django-sslserver==0.19
django-tz-detect==0.2.6
docutils==0.12
futures==3.0.5
jmespath==0.9.0
lazy-object-proxy==1.2.1
MySQL-python==1.2.5
oauthlib==1.0.3
pycrypto==2.6.1
PyJWT==1.4.0
pylint==1.5.4
pylint-django==0.7.1
pylint-plugin-utils==0.2.3
python-dateutil==2.5.0
python-linkedin==4.1
python-openid==2.2.5
python-social-auth==0.2.1
pytz==2016.3
PyYAML==3.11
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
South==1.0.2
wheel==0.26.0
wrapt==1.10.6
